+++ NEWS.md
@@ -2,6 +2,7 @@
 * `mallet_instance_list_store()` renamed to `instance_list_save()` and
 `mallet_instance_list_restore()` renamed to `instance_list_load()`.
 * `as_LDA()` outputs information on number of docs / alphabet size.
+* Method `svmlight()` moved here from package 'polmineR.misc'.
 
 # biglda v0.0.1